Refactor: GetTokensModal
